+++ zend/includes.h
@@ -45,6 +45,7 @@
 /**
  *  Include other files from this library
  */
+#include "../include/inivalue.h"
 #include "../include/ini.h"
 #include "../include/exception.h"
 #include "../include/streams.h"
